re-organization of unit-tests
