--- carina-api/src/test/java/com/qaprosoft/apitools/validation/mock/method/DeleteUserMethod.java
@@ -1,9 +1,0 @@
-package com.qaprosoft.apitools.validation.mock.method;
-
-import com.qaprosoft.carina.core.foundation.api.AbstractApiMethodV2;
-
-public class DeleteUserMethod extends AbstractApiMethodV2 {
-    public DeleteUserMethod() {
-        super();
-    }
-}
