+++ carina-api/src/test/java/com/qaprosoft/carina/core/foundation/api/AbstractApiMethodTest.java
@@ -15,26 +15,35 @@
  *******************************************************************************/
 package com.qaprosoft.carina.core.foundation.api;
 
-import com.qaprosoft.apitools.validation.mock.method.DeleteUserMethod;
-import com.qaprosoft.carina.core.foundation.api.ssl.PutDocMethod;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-public class AbstractApiMethodTest {
+import com.qaprosoft.carina.core.foundation.utils.Configuration;
+import com.qaprosoft.mock.apimethod.AutoReplaceUrlPartsMethod;
+import com.qaprosoft.mock.apimethod.NoPlaceholdersInURLMethod;
+import com.qaprosoft.mock.apimethod.PutDocMethod;
 
-    private final static String BODY_CONTENT = "{\"key\": \"value\"}";
-    private final static String EXPECTED_METHOD_PATH_URL = "https://jsonplaceholder.typicode.com/users/1";
+public class AbstractApiMethodTest {
 
     @Test
     public void testGetRequestBodyMethod() {
         PutDocMethod putDocMethod = new PutDocMethod();
-        putDocMethod.setBodyContent(BODY_CONTENT);
-        Assert.assertEquals(putDocMethod.getRequestBody(), BODY_CONTENT);
+        final String bodyContent = "{\"key\": \"value\"}";
+        putDocMethod.setBodyContent(bodyContent);
+        Assert.assertEquals(putDocMethod.getRequestBody(), bodyContent);
     }
 
     @Test
-    public void testInitMethod() {
-        DeleteUserMethod api = new DeleteUserMethod();
-        Assert.assertEquals(api.methodPath, EXPECTED_METHOD_PATH_URL);
+    public void testNoPlacehodlersInURL() {
+        NoPlaceholdersInURLMethod api = new NoPlaceholdersInURLMethod();
+        final String expectedMethodPath = "https://jsonplaceholder.typicode.com/users/1";
+        Assert.assertEquals(api.methodPath, expectedMethodPath);
+    }
+
+    @Test
+    public void testAutoReplacementInURL() {
+        AutoReplaceUrlPartsMethod method = new AutoReplaceUrlPartsMethod();
+        final String expectedMethodPath = Configuration.getEnvArg("base_url") + "/mock/part";
+        Assert.assertEquals(method.getMethodPath(), expectedMethodPath);
     }
 }
