+++ carina-api/src/test/java/com/qaprosoft/carina/core/foundation/api/log/CarinaAPILoggerTest.java
@@ -1,8 +1,17 @@
+package com.qaprosoft.carina.core.foundation.api.log;
+
+
+import java.lang.invoke.MethodHandles;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.testng.annotations.Test;
 
-import java.lang.invoke.MethodHandles;
+/**
+ * 
+ * @author Nikolay Zheleznui
+ *
+ */
 
 public class CarinaAPILoggerTest {
     private static final Logger LOGGER = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
