+++ carina-api/src/test/java/com/qaprosoft/carina/core/foundation/api/log/LoggingOutputStreamTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package com.qaprosoft.carina.core.foundation.log;
+package com.qaprosoft.carina.core.foundation.api.log;
 
 import java.io.IOException;
 import java.lang.invoke.MethodHandles;
