+++ carina-api/src/test/java/com/qaprosoft/mock/apimethod/NoPlaceholdersInURLMethod.java
@@ -1,0 +1,9 @@
+package com.qaprosoft.mock.apimethod;
+
+import com.qaprosoft.carina.core.foundation.api.AbstractApiMethodV2;
+
+public class NoPlaceholdersInURLMethod extends AbstractApiMethodV2 {
+    public NoPlaceholdersInURLMethod() {
+        super();
+    }
+}
