+++ carina-api/src/test/java/com/qaprosoft/mock/apimethod/XmlContentTypeMethod.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package com.qaprosoft.apitools.validation.mock.method;
+package com.qaprosoft.mock.apimethod;
 
 import com.qaprosoft.carina.core.foundation.api.AbstractApiMethodV2;
 import com.qaprosoft.carina.core.foundation.api.annotation.*;
